requirements.txt
quart
telethon
flask-httpauth
hypercornstreamlit
telethon
streamlit-autorefresh
pandas
Pillow
easyocr
numpy